% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoMat.R
\name{A}
\alias{A}
\title{Area of geographic intersection}
\usage{
A(mot, spdf)
}
\arguments{
\item{mot}{a notion}

\item{spdf}{Spatial Polygons Data Frame}
}
\value{
surface area contigency table of intersection between a notion and 
a spatial polygon data frame
}
\description{
Quantifies the area (in m2) of intersection between the geographical 
distribution of a notion and that of any other geographic subsection
}
\examples{
A("jardin",spdf.geo)

}
