% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoMat.R
\name{affichage}
\alias{affichage}
\title{Map of a notion}
\usage{
affichage(mot, partition)
}
\arguments{
\item{mot}{a notion}

\item{partition}{Spatial Polygons Data Frame}
}
\value{
None
}
\description{
Draws map of the different spellings for a notion
}
\examples{
affichage("jardin",spdf.geo)

}
