% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoMat.R
\name{affichage_clust}
\alias{affichage_clust}
\title{Dendrogram of lemma}
\usage{
affichage_clust(nom_de_la_fonction, mot, spdf.geo)
}
\arguments{
\item{nom_de_la_fonction}{name of algorithm to be used}

\item{mot}{a notion}

\item{spdf.geo}{Spatial Polygons Data Frame}
}
\value{
None
}
\description{
Draws a dendrogram with hierarchical clustering of lemmas
}
\examples{
affichage_clust("work in progress", "jardin",spdf.geo)

}
