% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoMat.R
\name{Caffiche}
\alias{Caffiche}
\title{Dendrogram cluster selection visual}
\usage{
Caffiche(Mots, Type_de_clust, Nb_groupe, parti)
}
\arguments{
\item{Mots}{a notion}

\item{Nb_groupe}{number of cluster}

\item{parti}{Spatial Polygons Data Frame}

\item{nom_de_la_fonction}{name of algorithm to be used}
}
\value{
None
}
\description{
Draws a dendrogram with hierarchical clustering of lemmas with box around
optimal clusters for given number of groups
}
\examples{
Caffiche("jardin","work in progress",3,spdf.geo)

}
