% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoMat.R
\name{calcul_ordre_dendro}
\alias{calcul_ordre_dendro}
\title{Order dendrogramme clustering}
\usage{
calcul_ordre_dendro(merge, n)
}
\arguments{
\item{merge}{work in progress}

\item{n}{number}
}
\value{
object
}
\description{
Finds the proper order for dendrogam
}
\examples{
Calcule_statistique_base("work in progress",3)

}
