% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoMat.R
\name{Calcule_statistique_base}
\alias{Calcule_statistique_base}
\title{Similarity statistics}
\usage{
Calcule_statistique_base(mot, spdf.geo)
}
\arguments{
\item{mot}{a notion}

\item{spdf.geo}{Spatial Polygons Data Frame}
}
\value{
list with notion and similarity statistics between the notion and 
a spatial partition. These include khi2, Cramer V2, specificity index,
location index and entropy.
}
\description{
Calculates smilarity statistics between a notion and a spatial partition
}
\examples{
Calcule_statistique_base("jardin",spdf.geo)

}
