% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoMat.R
\name{charger_Donnees}
\alias{charger_Donnees}
\title{Load files}
\usage{
charger_Donnees(cheminDonnees)
}
\arguments{
\item{cheminDonnees}{string of path to data}
}
\value{
None
}
\description{
Loads spatial data frames into memory
}
\examples{
charger_Donnees("data-raw")

}
