% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoMat.R
\name{clust_Entropie_notion}
\alias{clust_Entropie_notion}
\title{Clustering of notion by entropy}
\usage{
clust_Entropie_notion(mots, nbgroupe, spdf_geo)
}
\arguments{
\item{mots}{a notion}

\item{nbgroupe}{number of groups}

\item{spdf.geo}{Spatial Polygons Data Frame}
}
\value{
returns parameters for a dendrogram
}
\description{
hierarchical clustering of lemmas of a notion based on the the measure of 
entropy
}
\examples{
clust_Entropie_notion("jardin",3,spdf.geo)

}
