% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoMat.R
\name{clust_khi_geo}
\alias{clust_khi_geo}
\title{Clustering of geographical partition by khi2}
\usage{
clust_khi_geo(mots, nbgroupe, spdf_geo)
}
\arguments{
\item{mots}{a notion}

\item{nbgroupe}{number of groups}

\item{spdf.geo}{Spatial Polygons Data Frame}
}
\value{
returns parameters for a dendrogram
}
\description{
hierarchical clustering of geographical partition based on the the measure of 
khi2 statistic of independence
}
\examples{
clust_khi_geo("jardin",3,spdf.geo)

}
