% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoMat.R
\name{Entropie}
\alias{Entropie}
\title{Entropy}
\usage{
Entropie(A)
}
\arguments{
\item{A}{surface area contigency table of intersection between a notion and 
a spatial polygon data frame}
}
\value{
total entropy of given table
}
\description{
Calculates the total entropy of a contigency table
}
\examples{
Entropie(A("jardin",spdf.geo))

}
