% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoMat.R
\name{Entropie_ligne}
\alias{Entropie_ligne}
\title{Line Entropy}
\usage{
Entropie_ligne(A)
}
\arguments{
\item{A}{surface area contigency table of intersection between a notion and 
a spatial polygon data frame}
}
\value{
total entropy of given table
}
\description{
Calculates the total line entropy of a contigency table
}
\examples{
A = A("jardin",spdf.geo)
Entropie_ligne(A)

}
