% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{holidays}
\alias{holidays}
\title{French holidays dataset}
\format{
A data frame with 132 rows and 2 variables:
\describe{
  \item{ds}{date of holiday}
  \item{holiday}{Name of holiday}
}
}
\source{
\url{https://www.kaggle.com/ahmedlahlou/accidents-in-france-from-2005
-to-2016}
}
\usage{
holidays
}
\description{
A dataset containing the dates of the 11 main french holidays
from 2005 to 2016
}
\keyword{datasets}
