% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoMat.R
\name{inter_spdf}
\alias{inter_spdf}
\title{Geographic intersection}
\usage{
inter_spdf(spdf1, idx1, spdf2, idx2)
}
\arguments{
\item{spdf1}{first Spatial Polygons Data Frame}

\item{idx1}{column index of lemmas for spdf1}

\item{spdf2}{second Spatial Polygons Data Frame}

\item{idx2}{column index of lemmas for spdf2}
}
\value{
binary matrix of intersection between two spatial polygon data frame
}
\description{
indicates the presence (or lack) of intersection between polygons of a 
spatial dataframe
}
\examples{
inter_spdf(lspdf[[239]],2,spdf.geo,2)

}
