% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoMat.R
\name{map_clust}
\alias{map_clust}
\title{Sub dendrogram}
\usage{
map_clust(cut, spdf1)
}
\arguments{
\item{cut}{list of heights of hclust}

\item{spdf1}{Spatial Polygons Data Frame}
}
\value{
None
}
\description{
Draws dendogram with cut at optimal level. That is where there is the 
maximum height difference between two levels
}
\examples{
A("jardin",spdf.geo)

}
