% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{users}
\alias{users}
\title{Users in accident}
\format{
A data frame with 1876005 rows and 12 variables:
\describe{
  \item{Acc_number}{Accident identifier}
  \item{Num_Veh}{Identification of the vehicle taken back for each user
  occupying this vehicle (including pedestrians who are attached to the
  vehicles that hit them)}
  \item{place}{Allows to locate the place occupied in the vehicle by the
   user at the time of the accident}
  \item{catu}{User category:
   \itemize{
    \item{1 : Driver}
    \item{2 : Passenger}
    \item{3 : Pedestrian}
    \item{4 : Pedestrian in rollerblade or scooter}
   }
  }
  \item{grav}{everity of the accident: The injured users are classified into
  three categories of victims plus the uninjured
   \itemize{
    \item{Unscathed}
    \item{Killed}
    \item{Hospitalized wounded}
    \item{Light injury}
   }
  }
  \item{sex}{Sex of the user
   \itemize{
    \item{Male}
    \item{Female}
   }
  }
  \item{Year_on}{Year of birth of the user}
  \item{trip}{Reason for traveling at the time of the accident:
   \itemize{
    \item{1 : Home - work}
    \item{2 : Home - school}
    \item{3 : Shopping - Shopping}
    \item{4 : Professional use}
    \item{5 : Promenade - leisure}
    \item{9 : Other}
   }
  }
  \item{secu}{on 2 characters:
   \itemize{the first concerns the existence of a safety equipment
    \item{1 : Belt}
    \item{2 : Helmet}
    \item{3 : Children's device}
    \item{4 : Reflective equipment}
    \item{9 : Other}
   }
   \itemize{the second is the use of Safety Equipment
    \item{1 : Yes}
    \item{2 : No}
    \item{3 :  Not determinable}
    }
  }
  \item{locp}{Location of the pedestrian:
   \itemize{On pavement:
    \item{1 : More than 50 m from the pedestrian crossing}
    \item{2 : Less than 50 m from the pedestrian crossing}
   }
   \itemize{On pedestrian crossing:
    \item{3 : Without light signaling}
    \item{4 : With light signaling}
   }
   \itemize{Various:
    \item{5 : On the sidewalk}
    \item{6 : On the verge}
    \item{7 : On refuge or BAU}
    \item{8 : On against aisle}
   }
  }
  \item{actp}{Action of the pedestrian:
   \itemize{Moving
    \item{0 : not specified or not applicable}
    \item{1 : Meaning bumping vehicle}
    \item{2 : Opposite direction of the vehicle}
    }
   \itemize{Various
    \item{3 : Crossing}
    \item{4 : Masked}
    \item{5 : Playing - running}
    \item{6 : With animal}
    \item{9 : Other}
    }
  }
  \item{etatp}{This variable is used to specify whether the injured
  pedestrian was alone or not
   \itemize{
    \item{1 : Only}
    \item{2 : Accompanied}
    \item{3 : In a group}
   }
  }

}
}
\source{
\url{https://www.kaggle.com/ahmedlahlou/accidents-in-france-from-2005
-to-2016}
}
\usage{
users
}
\description{
A dataset with details about the users implicated in the accidents logged in
the dataset caracteristics and places
}
\keyword{datasets}
