% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vehicles}
\alias{vehicles}
\title{Vehicles in accident}
\format{
A data frame with 1433389 rows and 9 variables:
\describe{
  \item{Num_Acc}{Accident ID}
  \item{Num_Veh}{Identification of the vehicle taken back for each user
  occupying this vehicle (including pedestrians who are attached to vehicles
  that hit them) - alphanumeric code}
  \item{GP}{
   \itemize{
    \item{1 : PK or PR or increasing postal address number}
    \item{2 : PK or PR or descending postal address number}
   }}
  \item{CATV}{Category of vehicle:
   \itemize{
    \item{Bicycle}
    \item{Moped <50cm3}
    \item{Cart (Quadricycle with bodied motor) (formerly "cart or motor
    tricycle")}
    \item{Not used since 2006 (registered scooter)}
    \item{Not used since 2006 (motorcycle)}
    \item{Not used since 2006 (side-car)}
    \item{VL only}
    \item{Not used category (VL + caravan)}
    \item{Not used category (VL + trailer)}
    \item{VU only 1,5T <= GVW <= 3,5T with or without trailer (formerly VU
    only 1,5T <= GVW <= 3,5T)}
    \item{Most used since 2006 (VU (10) + caravan)}
    \item{Most used since 2006 (VU (10) + trailer)}
    \item{PL only 3,5T}
   }
  }
}
}
\source{
\url{https://www.kaggle.com/ahmedlahlou/accidents-in-france-from-2005
-to-2016}
}
\usage{
vehicles
}
\description{
A dataset with details about the vehicles implicated in the accidents
logged in the dataset caracteristics and places
}
\keyword{datasets}
